Create basic CLI template
When creating CLI tools it's nice to be able to gather the tools into a
single CLI. However, one does not want to rewrite all that needs to be
done to create the full CLI. Such scaffolding could be, dicovery of
commands, coherent help texts, sharing config etc.

This commit would provide a simple framework where commands can be added
by just providing an executable and config can be read into the
environment and shared between the commands of the tool with a single
file.

To a developer this should make it fairly simple to create a CLI tool
around multiple scripts and create a cohesive experience.