Add subcommands with completions
Without completions we won't get any help to remember what commands the
cli tool takes, which makes it more difficult to use. We want to be able
to create at tree of commands and subcommands that should all be
possible to complete in bash.

We want to be able to either pick up on subcommand completions
automatically or specifically define the subcommand completions in
files. We therefore allow for two strategies:

1. Either you create a file tree which matches the command/subcommand
   structure and the bash completion script will find the available
   commands. The structure would look like:

	|-- cli.sh
	|-- commands/
	|     |-- one.sh
	|     |-- one-commands/
	|     |     |-- two.sh
	|     |     |-- three.sh

   where 'two.sh' and 'three.sh' are sub-commands to one and would be
   found if the user types 'cli one {tab}{tab}'.

2. Or, you can create a textfile listing the available commands at a
   specific piont in the tree. Each subcommand would have its own
   completion list file, and each file needs to be named using a
   specific format. To achive the same as in 1. one would create:

	|-- cli.sh
	|-- commands/
	|     |-- one.sh
	|     |-- one-completions.txt

   where 'one-completions.txt' would contain:

	two
	three

   that is, one subcommand on each line.

one could combine 1 and 2 as well as creating deeper structures of
completion hint directories and files.