Stop overwriting partial flag
When writing 'cli direction -e<tab>' the completion script overwrites
'-e' with 'direction' so the string auto complets to:

cli direction direction

What should happen is that the '<tab>' press should not touch the
string. Rather, it should suggest the sub-commands to 'direction'.

The reason this happens is that we filter away the '-e' and do not add
anything to the end of the array of filtered words. This means that the
array of filtered words looks like:

('cli' 'direction')

And since the last entry of COMP_WORDS is always an empty string ('')
the subsequent logic assumes that only 'cli' has been entered and that
we have written 'direction' so it auto completes to:

'cli direction direction'

The solution we choose is to add the empty string at the end back so the
completion script knows that we have written 'cli direction' and
nothing more so it can fetch all the sub-command completions for the
'direction' command.