Stop send completion hint to config finding func
At the moment we are sending the whole 'COMP_WORDS' array to the config
finding function. This means that the function the logic after the
function needs to deal with partial commands.

When typing 'cli direction do<tab><tab>' the filtered list of words
looks like:

('cli' 'direction' 'do')

this array is sent to the config finding function, which then has to
deal with 'do' by searching backwards in the list if it doesn't find a
file. This also results in us having to search for directories and
config files by removing and appending words at the end of the search
attempt.

We note that When typing 'cli direction down <tab><tab>' the filtered
list of words (as we are mimicing 'COMP_WORDS') looks like:

('cli' 'direction' 'down' '')

This empty string is always present once we have provided a full
sub-command. The last entry of the 'COMP_WORDS' array is the 'completion
hint' (our name for the last entry). This can be used by 'compgen' to
filter among the suggestions and potentially auto-complete if there is
only one suggestion.

For our purposes, however, it should not be used to find the completion
suggestions. In our case we need the everything but the completion hint.
So this means that we can always skip the last item in the array if we
want to get the list of the finished command.