Allow config files for base command
When one writes 'cli <tab>' the completion file cannot find the flag
completion config for the base command 'cli'. This is an issue as we
will most likely have flags for the base command that we want to be able
to suggest as hints. Therefore the helper function for finding
completion configs need to support this case as well.

We want the helper function to work for the base command both when
finding flags and when finding sub commands. This presents a challenge
as we need to define a "state" in the function for when to look for
completion configs for the base command. This state is when the input to
the function looks like:

('cli' '')

or

('cli' '<some-string>')

as this is the two cases where we expect the completion suggestions to
be relevant for the 'cli' command. The issue is that for the normal
case, i.e. for regular sub-commands in '<command>-commands' directories
and '<command>-completions.txt' files, we need to remove the
'<some-string>' input as this messes with the tree traversal. It messes
with it because for this input:

('cli' 'sub-command' '<some-string>')

the helper function should look for completion configs for the
'sub-command', but if it got the '<some-string>' item it would look for
completion configs relative to this string.

The solution is to make sure that we always send an empty item as the
last item of the input to the function. That way we can always detect
the states for when to look for configs for the base command and when to
look for general configs.