Implement flag completion, flags with no arguments
When writing 'cli -<tab><tab>' or 'cli direction -<tab><tab>' we get
the subcommands as completions. Instead we want to get the possible
flags that we can provide to the command. Similarly, if we write
'cli <tab><tab>' or 'cli dig <tab><tab>' we want the available flags to
be present in the completion list. At the moment there are no flags in
the completion list and we don't filter out the sub commands from the
completion suggestion to only show the flags.

What we want is to provide a completion file, similar to sub-command
discovery, where we specify the command. So in the same vein as with
sub-commands, we'll allow the use of a text file, listing the available
flags for a command. Each command would again have their own flag
completion file. An example of this would be:

	|-- cli.sh
	|-- commands/
	|     |-- one.sh
	|     |-- one-flags.txt
	|     |-- one-commands/
	|     |     |-- three.sh
	|     |     |-- two.sh
	|     |     |-- two-flags.txt

where 'one-flags.txt' could contain:

	-h
	-w

and 'two-flags.txt' could contain:

	-x
	-q

One would then get the following completions:

	cli one {tab}{tab}
	# -h    three two   -w
	cli one -{tab}{tab}
	# -h -w
	cli one -w {tab}{tab}
	# -h    three two
	cli one -w two {tab}{tab}
	# -q -x