REFACTOR: Handle all number of commands same way
We are handling the case of the first level commands in a separate
execution path to the deeper level commands. Level of commands:

	cli {first-level} {second-level} {third-level} ...

However, the only difference between the two cases is where
the commands directory is located:

	|-- cli.sh
	|-- commands/
	|     |-- one.sh
	|     |-- one-commands/
	|     |     |-- two.sh
	|     |     |-- two-commands/

with the first level commands residing under the 'commands' directory
and the deeper level existing in '{command-name}-commands' directories.

These should be possible to handle in the same way to make the logic a
little bit more easy to reason around and understand (less branching
should hopefully be easier to read). This commit will hopefully simplify
a little bit.