REFACTOR: Simplify logic for input parsing
The parsing of the input, where we determine what flags, commands and
completion hint has been provided is more complex than it needs to be.
We really just need to pick the last item of the input array and then
split the rest into flags and commands.

With that simplification we can also remove some of the workaround used
when finding completion files and directories. All in all the completion
script should get much more easy to follow.