Allow flags that takes arguments
At the moment flags are assumed to not take any arguments. This means
that for an argument '-t' that takes an argument the auto completion is
broken. Writing 'cli -t type {tab}{tab}' does not show any completions
(given that 'type' is not a command for 'cli'). As it is very common to
have flags that takes commands we need to allow this.

The solution in this commit is to allow for a hint in the flag
completions file:

	-e
	-t {TYPE}

Where '{TYPE}' could be any string. This solution uses the 'COMPREPLY'
variable to show the hint. The side effect of this is that the
formatting becomes a little wonky. As the content of 'COMPREPLY' will be
printed with an indentation that has the same width as the content. So a
flag hint like 'lorem ipsum dolor sit amet' would look something like
this:

	cli -f {tab}{tab}
	                            lorem ipsum dolor sit amet

So we should recommend that the hints are kept short. There are
potentially other ways to show the hint, one theory was that we should
be able to use ANSI escape codes [1] to move the cursor around together
with 'printf' to print the hint on the next line and then return the
cursor to the previous position. Something like:

	printf "\0337\n\r%s\0338\033[1A" "$flag_argument_hint"

Where the position on the line is saved, the hint is printed on a new
line and we then return to the correct prosition and up one line. It
didn't quite work, either we got the hint printed stacked below like:

	cli -t {tab}{tab}{tab}{tab}
	{TYPE}
	{TYPE}
	{TYPE}

or the cursor kept moving upward, like so (cursor illustrated by []):

	       []
	       []
	cli -t []{tab}{tab}{tab}{tab}
	{TYPE}

Where the cursor would be in the right position after the first
completion but with repeated completion {tab} presses the cursor would
climb upwards.

As the use of 'COMPREPLY' seems to work decently we'll leave the pretty
formatting for another day.

Refs:
[1] https://en.wikipedia.org/wiki/ANSI_escape_code