Make auto completion zsh & mac compatible
Trying to run the autocompletion in a zsh shell does not provide you
with auto completions. This seems to be because either zsh handles
arrays differently to bash [1], this shouldn't be the case as long as
the shebang is #!/usr/bin, or because of how zsh handles the
'COMP_WORDS' array (have not found a reason to believe so).

In practice, what we see is that when running in zsh there are extra
empty items in the arrays, so when filtering or splitting lists where we
"know" that there shouldn't be any empty items we don't get the right
result. Filtering out these empty "ghost" items works, and do not impact
funcionality in bash.

There is also a difference in how auto completion works when it comes to
selecting completion suggestions, between bash and zsh. In bash we can
"trick" the auto completion to not select a "single" suggestion by
adding an empty string, "", as a second suggestion. In zsh this does not
work and the non-empty suggestion is selected automatically. We use this
trick in bash to show the hints for flags, this does not work in zsh.
Instead we'll show no completion until the user has provided an argument
to the flag, hinting that the command is waiting for more, which at
least is something.

Additionally, the implemenation of find on MacOS does not support the
'-executable' option. Instead we need to use the more POSIX-compliant
'-perm' (or permissions)  option. Additoinally there are multiple ways
of specifying permissions, but not all of them are portable [2]. Testing
on Linux and MacOS gives that '-perm -111' seems to be portable enough
for our purposes and give an acceptable result.

Refs:
[1] https://apple.stackexchange.com/questions/361870/what-are-the-practical-differences-between-bash-and-zsh
[2] https://stackoverflow.com/a/29039160