Describe how to use custom base script
Users may not want to use the 'cli.sh' script when creating their own
CLI tool. It may be that they need other functionality or cannot work
within the constraints of the 'cli.sh' script. Is is very doable to only
use the framework in order to get auto completion, so we want to
describe this in order to enable this workflow.