Allow commands to be links to executables
An obstacle for adoption is that ShellFrame needs the executables to be
in the 'commands' directory. Often tools may reside in different places
on the system and in different repos. Users, most likely, don't want to
move all their stuff to one place just to accomodate ShellFrame. A user
could solve this by writing a wrapper script that is placed in the
'commands' directory and calls the original script in its original
place. However, users may find it easier to just create links to their
tools instead.