Auto discover scripts in commands directory
If users have working scripts that they want to add to a tool based on
ShellFrame they need to write wrapper scripts or update the scripts to
match ShellFrame's structure.

Additionally one often have a group of related scripts that it would be
nice to more or less dump into a directory and have them work with a cli
tool.

This commit enables the cli.sh to find executable scripts in a similar
way to the auto completion script finds completion files. If a command
like the following is given:

    cli foo bar baz

the cli script will look for and execute the following path:

    ./commands/foo-commands/bar-commands/baz.<EXTENSION>

where the 'baz.<EXTENSION>' file must be executable. Options, sub
commands and other input will be passed to the 'baz' script.

Help-texts for commands like this are enabled by allowing users to
provide a '<COMMAND>-help.txt' file which will be printed similarly to
the other help texts when 'cli -h' is invoked.

This way scripts can be grouped as commands under foo (and/or bar) by
just copy and pasting (or linking) existing scripts the commands
directory like so:

            |-- cli.sh
            |-- commands/
            |     |-- foo-help.txt
            |     |-- foo-completions.txt
            |     |-- foo-commands/
            |     |     |-- hey.sh
            |     |     |-- hoo.js
            |     |     |-- bar-completions.txt
            |     |     |-- bar-help.txt
            |     |     |-- bar-commands/
            |     |     |     |-- qux
            |     |     |     |-- flux.sh
            |     |     |     |-- baz.py

The '-h' option for the base command 'cli.sh', will only find the first
level of '<COMMAND>-help.txt' files, so 'foo-help.txt' in the example
above. The 'cli.sh' script will be able to find the other
'<COMMAND>-help.txt' if requested explicitly, for example:

    cli foo -h
    cli foo bar -h

Scripts still have to accept the '-v' option (as in not break if '-v' is
passed) as this option is passed along to the final command. This seems
like a small update to most scripts, though it could be troublesome if
'-v' already has some other meaning for the script. But that seems like
a minor issue at this point.