Move cli.sh help to external help file
The help-text of cli.sh is hard coded in the script file. If a user
wants to copy paste the cli.sh file and use it for their own cli they
might update the help text. If they later want to update the cli.sh with
a new version from this repo they would have to edit the cli.sh again
rather than only copy pasting the cli.sh.

This commit creates a cli-help.txt file which will hold the help text
that will be printed if the '-h' text is provided to cli.sh. It also
creates a 'cli-usage.txt' that the cli.sh will read any time it needs to
print the usage instructions.